fix stuff, make things nicer
