+++ src/ProductConstructionService/ProductConstructionService.Api/appsettings.json
@@ -13,7 +13,7 @@
     },
     "AllowedHosts": "*",
     "DefaultWorkItemQueueName": "pcs-workitems",
-    "DefaultWorkItemConsumerCount": 1,
+    "DefaultWorkItemConsumerCount": 4,
     "CodeFlowWorkItemQueueName": "pcs-codeflow-workitems",
     "WorkItemConsumerOptions": {
         "QueuePollTimeout": "00:01:00",
